pyproject.toml
[project]
name = "linz-s3-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Quinn Hornblow", email = "quinn.hornblow@protonmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pystac>=1.14.3",
    "pystac-client>=0.9.0",
    "requests-cache>=1.3.1",
    "shapely>=2.1.2",
    "odc-stac>=0.5.2",
    "odc-geo>=0.5.1",
    "pydantic>=2.13.3",
    "tqdm>=4.67.3",
]

[build-system]
requires = ["uv_build>=0.10.11,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "deptry>=0.25.1",
    "notebook>=7.5.5",
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]

[tool.deptry]
known_first_party = ["linz_s3_utils"]
